move findPath to GameMap; findPath tests TC26-TC33 pass
